+++ packages/seroval/src/core/tree/web-api.ts
@@ -7,52 +7,8 @@ import { serializeArrayBuffer } from './primitives';
 import type {
   SerovalBlobNode,
   SerovalFileNode,
-  SerovalURLNode,
-  SerovalURLSearchParamsNode,
 } from '../types';
 import { SerovalNodeType } from '../constants';
-
-export function createURLNode(
-  ctx: ParserContext,
-  id: number,
-  current: URL,
-): SerovalURLNode {
-  assert(ctx.features & Feature.WebAPI, new UnsupportedTypeError(current));
-  return {
-    t: SerovalNodeType.URL,
-    i: id,
-    s: serializeString(current.href),
-    l: undefined,
-    c: undefined,
-    m: undefined,
-    d: undefined,
-    f: undefined,
-    a: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
-
-export function createURLSearchParamsNode(
-  ctx: ParserContext,
-  id: number,
-  current: URLSearchParams,
-): SerovalURLSearchParamsNode {
-  assert(ctx.features & Feature.WebAPI, new UnsupportedTypeError(current));
-  return {
-    t: SerovalNodeType.URLSearchParams,
-    i: id,
-    s: serializeString(current.toString()),
-    l: undefined,
-    c: undefined,
-    m: undefined,
-    d: undefined,
-    f: undefined,
-    a: undefined,
-    b: undefined,
-    o: undefined,
-  };
-}
 
 export async function createBlobNode(
   ctx: ParserContext,
